Add files via upload
This repository contains my work from the first week of sheCodes Workshops. It has been a fantastic learning experience with plenty of hands-on practice. Throughout this week, I delved into various coding concepts and applied them in practical scenarios. I'm excited to share my progress and look forward to the upcoming weeks of learning and development.